extract role options to data module
