+++ src/data/index.js
@@ -1,3 +1,4 @@
 export { default as levels } from './levels';
 export { default as manager } from './manager';
+export { default as roleOptions } from './roleOptions';
 export { default as staff } from './staff';
