+++ src/data/roleOptions.js
@@ -1,0 +1,12 @@
+const roleOptions = [
+  {
+    displayText: 'Manager',
+    optValue: 'manager',
+  },
+  {
+    displayText: 'Staff',
+    optValue: 'staff',
+  }
+];
+
+export default roleOptions;
